% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassWateRdash-package.R
\docType{package}
\name{MassWateRdash-package}
\alias{MassWateRdash}
\alias{MassWateRdash-package}
\title{MassWateRdash: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Marcus Beck \email{mbeck@tbep.org} (\href{https://orcid.org/0000-0002-4996-0059}{ORCID})

Authors:
\itemize{
  \item Marcus Beck \email{mbeck@tbep.org} (\href{https://orcid.org/0000-0002-4996-0059}{ORCID})
  \item Ben Wetherill \email{bwetherill@acasak.com} (\href{https://orcid.org/0000-0002-8791-4256}{ORCID})
  \item Mariel Sorlien \email{msorlien@nbep.org} (\href{https://orcid.org/0000-0001-7102-2918}{ORCID})
  \item Christopher Whitney \email{Chris.Whitney@unh.edu} (\href{https://orcid.org/0000-0003-2349-9134}{ORCID})
  \item Janelle Goeke \email{janelle.goeke@maine.edu} (\href{https://orcid.org/0000-0001-8056-4540}{ORCID})
  \item Jill Carr \email{Jillian.Carr@umb.edu} (\href{https://orcid.org/0000-0003-1476-6640}{ORCID})
}

}
\keyword{internal}
